Move GradleDistributionType to public package
